// webclient/src/api/generated/dtos/tickerResultSourcesItem.gen.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Sentiment Search API
 * Core API for handling user searches and delegating tracking tasks.
 * OpenAPI spec version: 1.0.0
 */

export type TickerResultSourcesItem = {
  url: string;
  snippet: string;
  /**
     * Timestamp of the last update moment in seconds since Unix epoch
     * @minimum 0
     */
  updatedAtSec: number;
  /**
     * Timestamp of the scrape moment in seconds since Unix epoch
     * @minimum 0
     */
  scrapedAtSec: number;
} & {
  /**
     * @minimum -1
     * @maximum 1
     */
  score: number;
};